+++ callbacks.go
@@ -1,8 +1,9 @@
 package alpm
 
 /*
+#include <stdint.h>
 #include <alpm.h>
-void logCallback(unsigned int level, char *cstring);
+void logCallback(uint16_t level, char *cstring);
 void go_alpm_log_cb(alpm_loglevel_t level, const char *fmt, va_list arg);
 void go_alpm_set_logging(alpm_handle_t *handle);
 */
@@ -10,7 +11,7 @@ import "C"
 
 var DefaultLogLevel = LogWarning
 
-func DefaultLogCallback(lvl uint, s string) {
+func DefaultLogCallback(lvl uint16, s string) {
 	if lvl <= DefaultLogLevel {
 		print("go-alpm: ", s)
 	}
@@ -19,11 +20,11 @@ func DefaultLogCallback(lvl uint, s string) {
 var log_callback = DefaultLogCallback
 
 //export logCallback
-func logCallback(level uint, cstring *C.char) {
+func logCallback(level uint16, cstring *C.char) {
 	log_callback(level, C.GoString(cstring))
 }
 
-func (h *Handle) SetLogCallback(cb func(uint, string)) {
+func (h *Handle) SetLogCallback(cb func(uint16, string)) {
 	log_callback = cb
 	C.go_alpm_set_logging(h.ptr)
 }
